small improvements

+++ horde/g_horde.h
@@ -5,7 +5,7 @@ void Horde_Init()  ;
 void Horde_RunFrame()  ;
 void ResetGame()  ;
 void HandleResetEvent()  ;
-extern int last_wave_number;  // Tracks the last completed wave number, used for intermission
+extern uint64_t last_wave_number;  // Tracks the last completed wave number, used for intermission
 
 // Item selection in Horde mode
 gitem_t* G_HordePickItem()  ;
@@ -34,6 +34,6 @@ extern gtime_t horde_message_end_time;  // Add this line
 extern void ClearHordeMessage();
 extern void UpdateHordeMessage(std::string_view message, gtime_t duration);
 extern void UpdateHordeHUD();
-extern int32_t g_totalMonstersInWave;
+extern uint32_t g_totalMonstersInWave;
 
 extern void CleanupInvalidEntities();
